refactor(Validator): refactor mvc pattern
